+++ info.h
@@ -8,11 +8,11 @@ struct Info
 {
 	D3DMATERIAL9	Material;
 	BOOL			isEnter;	// 入り込む
-	float			Distance;	// 交点までの距離
+	double			Distance;	// 交点までの距離
 	sp				Cross;		// 交点座標
 	sp				Vertical;	// 法線ベクトル
 	const Node *	pNode;		// 
-	float	Refractive;			// 屈折率
+	double	Refractive;			// 屈折率
 	Info()
 	{
 		pNode = NULL;
