+++ plane.cpp
@@ -10,7 +10,7 @@ BOOL	Plane::IsInside( const sp& L ) const
 
 BOOL Plane::GetInfo(const sp& K, const sp& L, Info& info) const
 {
-	float	t = ( K.z ) ? -L.z / K.z : (( L.z > 0 ) ? DBL_MAX : -DBL_MAX);
+	double	t = ( K.z ) ? -L.z / K.z : (( L.z > 0 ) ? DBL_MAX : -DBL_MAX);
 
 	if ( t <= 0 )
 		return FALSE;
