+++ rotate.h
@@ -5,9 +5,9 @@
 
 struct	rotate : public matrix
 {
-	rotate( sp axis, float angle);
-	rotate( float nx = 0.0, float ny = 0.0, float nz = 1.0, float angle = 0.0 );
-	void	init( float nx, float ny, float nz, float angle );
+	rotate( sp axis, double angle);
+	rotate( double nx = 0.0, double ny = 0.0, double nz = 1.0, double angle = 0.0 );
+	void	init( double nx, double ny, double nz, double angle );
 };
 
 #endif
